# src/openapi.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AlexMubarakshin/ApiSchema/1.0.1
info:
  description: Api schema
  version: "1.0.1"
  title: API schema
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/dialogs:
    get:
      tags:
        - dialogs
      description: "Получение списка диалогов пользователя"
      summary: "Получить список диалогов"
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          description: "Строка для поиска диалога"
          required: false
          schema:
            type: string
            example: "Something"
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: "Успешное получение списка Диалогов"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Dialog'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - dialogs
      description: "Создание диалога"
      summary: "Создать диалог"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DialogUserShortList"
      responses:
        "200":
          description: "Успешное создание Диалога"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: number
                            description: "Id созданного диалога"
                            example: 45
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/dialogs/unreaded:
    get:
      tags:
        - dialogs
      description: "Получение общего кол-ва непрочитанных сообщений"
      summary: "Получение общего кол-ва непрочитанных сообщений"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "Успешное получение общего кол-ва непрочитанных сообщений"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          count:
                            type: number
                            example: 1000
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/dialogs/{id}:
    delete:
      tags:
        - dialogs
      description: "Удаление диалога"
      summary: "Удалить диалог"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: "Id диалога"
          required: true
          schema:
            type: number
            example: 53
      responses:
        "200":
          description: "Успешное удаление Диалога"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: number
                            description: "Id удаленного диалога"
                            example: 45
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/dialogs/{id}/messages:
    get:
      tags:
        - dialogs
      description: "Получение списка сообщений в диалоге"
      summary: "Получить список сообщений в диалоге"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: "Id диалога"
          required: true
          schema:
            type: number
            example: 53
        - name: query
          in: query
          description: "Строка для поиска сообщения"
          required: false
          schema:
            type: string
            example: "Something"
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: "Успешное получение списка сообщений"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - dialogs
      description: "Отправка сообщений"
      summary: "Отправка сообщений"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: "Id диалога"
          required: true
          schema:
            type: number
            example: 53
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeasageSendRequestBody"
      responses:
        "200":
          description: "Успешная отправка сообщения"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Message"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/dialogs/{dialog_id}/messages/{message_id}:
    delete:
      tags:
        - dialogs
      description: "Удаление сообщения"
      summary: "Удаление сообщения"
      security:
        - bearerAuth: []
      parameters:
        - name: dialog_id
          in: path
          description: "Id диалога"
          required: true
          schema:
            type: number
            example: 53
        - name: message_id
          in: path
          description: "Id сообщения"
          required: true
          schema:
            type: number
            example: 1141
      responses:
        "200":
          description: "Успешное удаление сообщения"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message_id:
                            type: number
                            description: "Id удаленного сообщения"
                            example: 1141
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - dialogs
      description: "Редактирование сообщения"
      summary: "Редактирование сообщения"
      security:
        - bearerAuth: []
      parameters:
        - name: dialog_id
          in: path
          description: "Id диалога"
          required: true
          schema:
            type: number
            example: 53
        - name: message_id
          in: path
          description: "Id сообщения"
          required: true
          schema:
            type: number
            example: 1141
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeasageSendRequestBody"
      responses:
        "200":
          description: "Успешное редактирование сообщения"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Message"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/dialogs/{dialog_id}/messages/{message_id}/recover:
    put:
      tags:
        - dialogs
      description: "Восстановление удаленного сообщения"
      summary: "Восстановление удаленного сообщения"
      security:
        - bearerAuth: []
      parameters:
        - name: dialog_id
          in: path
          description: "Id диалога"
          required: true
          schema:
            type: number
            example: 53
        - name: message_id
          in: path
          description: "Id сообщения"
          required: true
          schema:
            type: number
            example: 1141
      responses:
        "200":
          description: "Успешное восстановление сообщения"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Message"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/dialogs/{dialog_id}/messages/{message_id}/read:
    put:
      tags:
        - dialogs
      description: "Пометить сообщение как \"Прочитанное\""
      summary: "Пометить сообщение как \"Прочитанное\""
      security:
        - bearerAuth: []
      parameters:
        - name: dialog_id
          in: path
          description: "Id диалога"
          required: true
          schema:
            type: number
            example: 53
        - name: message_id
          in: path
          description: "Id сообщения"
          required: true
          schema:
            type: number
            example: 1141
      responses:
        "200":
          description: "Успешное прочтение сообщения"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "ok"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Тип ошибки"
          example: "invalid_request"
          enum:
            - invalid_request
            - unauthorized
        error_description:
          type: string
          items:
            $ref: '#/components/schemas/ErrorDescription'
    ErrorDescription:
      type: object
      properties:
        error_description:
          type: string
          description: "Error description"
          enum:
            - "Unauthorized"
            - "An authorization code must be supplied"
            - "Redirect URI mismatch"
            - "Invalid authorization code: CODE"
            - "Bad credentials"
    Response:
      type: object
      properties:
        error:
          type: string
          description: "Сообщение, в случае ошибки здесь будет причина"
        timestamp:
          type: number
          example: 1559751301818
    ListResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            total:
              type: integer
              description: "Количество элементов в списке"
              format: int64
            offset:
              type: integer
              description: "Отступ от начала списка"
              format: int64
            perPage:
              type: integer
              description: "Количество элементов на страницу"
              format: int64
              default: 20
...
    Dialog:
      type: object
      description: "Диалог пользователя"
      properties:
        id:
          type: number
          example: 10
        unread_count:
          type: number
          description: "Число непрочитанных сообщений в диалоге"
          example: 0
        last_message:
          $ref: "#/components/schemas/Message"
    DialogUserShortList:
      type: object
      properties:
        user_ids:
          type: array
          description: "Id пользователей, c которыми создается диалог"
          items:
            type: number
            description: "Id пользователя"
            example: 3
    Message:
      type: object
      description: "Личное сообщение"
      properties:
        id:
          type: number
          example: 12
        time:
          type: number
          description: "Дата и время отправки"
          format: date-time
          example: 1559751301818
        author_id:
          type: number
          example: 1
        recipient_id:
          type: number
          example: 1
        message_text:
          type: string
        read_status:
          $ref: "#/components/schemas/ReadStatus"
    MeasageSendRequestBody:
      type: object
      description: "Личное сообщение"
      properties:
        message_text:
          type: string
...
    ReadStatus:
      type: string
      description: "Статус прочтения"
      enum:
        - SENT
        - READ
...
  parameters:
    offsetParam:
      in: query
      description: "Отступ от начала списка"
      name: offset
      required: false
      schema:
        type: integer
        format: int64
    limitParam:
      in: query
      description: "Количество элементов на страницу"
      name: itemPerPage
      required: false
      schema:
        default: 20
        type: integer
        format: int64